renamed fed.py to base.py and added new files for other categories of data
